requirements.txt
# Core Kafka & Database
kafka-python-ng==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing & Analytics
pandas==2.1.4
numpy==1.26.2

# Machine Learning & Anomaly Detection
scikit-learn==1.3.2
joblib==1.3.2

# Data Generation
faker==19.3.1

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dateutil==2.8.2